+++ src/utils/promptUtils.ts
@@ -72,3 +72,19 @@ export async function promptValues(structureVariables: Variable[], structureOpti
 
     return { variables, optionals };
 }
+
+export async function promptShowSkippedItems(skippedItems: { [key: string]: string }, createdItemsCount: number) {
+    if (skippedItems) {
+        const showSkippedItems = await vscode.window.showInformationMessage(`Successfully created ${createdItemsCount} items and skipped ${Object.keys(skippedItems).length} items.`, 'Show overview', 'OK') === 'Show overview';
+
+        if (showSkippedItems) {
+            let skippedItemsOverview: string = 'Skipped items:';
+            for (const item in skippedItems) {
+                skippedItemsOverview += `\n\n- ${item}:\n      ${skippedItems[item]}`;
+            }
+            await vscode.window.showInformationMessage(skippedItemsOverview, { modal: true });
+        }
+    } else {
+        vscode.window.showInformationMessage(`Successfully created ${createdItemsCount} items.`);
+    }
+}